v3/pyproject.toml
[project]
name = "v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "numpy-quaternion>=2024.0.13",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "snoop>=0.6.0",
    "xgboost>=3.2.0",
]
